+str #17461 add a parallel version of foreach
